gravel: setting hostname must update state
We need to update the in-memory state for the node setting a new
hostname; otherwise we will be joining or creating a new cluster with,
potentially, the old hostname.
