+++ src/glass/src/app/shared/services/api/nodes.service.ts
@@ -66,6 +66,10 @@ export type DiskSolution = {
   possible: boolean;
 };
 
+export type SetHostnameRequest = {
+  name: string;
+};
+
 @Injectable({
   providedIn: 'root'
 })
@@ -115,5 +119,13 @@ export class NodesService {
    */
   deploymentDiskSolution(): Observable<DiskSolution> {
     return this.http.get<DiskSolution>(`${this.deploymentURL}/disksolution`);
+  }
+
+  /**
+   * Setup hostname
+   */
+   setHostname(name: string): Observable<boolean> {
+    const request: SetHostnameRequest = { name };
+    return this.http.put<boolean>(`${this.url}/hostname`, request);
   }
 }
