+++ src/glass/src/app/shared/services/api/orch.service.ts
@@ -39,10 +39,6 @@ export type HostDevices = {
   devices: Device[];
 };
 
-export type SetHostnameRequest = {
-  name: string;
-};
-
 @Injectable({
   providedIn: 'root'
 })
@@ -65,11 +61,4 @@ export class OrchService {
     return this.http.get<{ [hostName: string]: HostDevices }>(`${this.url}/devices`);
   }
 
-  /**
-   * Setup hostname
-   */
-  setHostname(name: string): Observable<boolean> {
-    const request: SetHostnameRequest = { name };
-    return this.http.put<boolean>(`${this.url}/hostname`, request);
-  }
 }
