+++ src/gravel/api/nodes.py
@@ -81,6 +81,10 @@ class StartDeploymentRequest(BaseModel):
     ntpaddr: str = Field(title="NTP server address")
 
 
+class SetHostnameRequest(BaseModel):
+    name: str = Field(min_length=1, title="The system hostname")
+
+
 @router.get("/deployment/disksolution", response_model=DiskSolution)
 async def node_get_disk_solution(request: Request) -> DiskSolution:
     """
@@ -215,6 +219,20 @@ async def nodes_get_token(request: Request):
     return TokenReplyModel(
         token=(token if token is not None else "")
     )
+
+
+@router.put("/hostname", response_model=bool)
+async def put_hostname(request: Request, req: SetHostnameRequest) -> bool:
+    nodemgr: NodeMgr = request.app.state.nodemgr
+    if nodemgr.deployment_state.deployed or nodemgr.deployment_state.ready:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+                            detail="Node already deployed")
+    try:
+        await nodemgr.set_hostname(req.name)
+    except Exception as e:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+                            detail=str(e))
+    return True
 
 
 router.add_websocket_route(  # pyright: reportUnknownMemberType=false
